Solve Issue#14 && error message if word cut

+++ src/jdis/jdis.h
@@ -25,7 +25,7 @@ extern double jdis(bst *p, bst *q, size_t card_interction);
 //  renvoie l'avl associé, en cas d'erreur retourne nullptr et désalloue toute
 //  zone mémoire précédemment allouée, si le nom de fichier ou le fourre-tout
 //  sont des pointeurs null, retourne nullptr.
-extern bst *file_to_bst(char *file_name, int value_max, bool wp,
+extern bst *file_to_bst(char *file_name, size_t value_max, bool wp,
     bst *uni_words);
 
 // add_element : Renvoie une valeur nulle si l'ajout en bout de chemin de ref
